# contracts/catalog.openapi.yaml
openapi: 3.1.0
info:
  title: deedlit.catalog
  version: 0.1.0
  description: >
    Source-of-truth catalog over Postgres + RustFS. Images, tags, params,
    references, ratings, favorites, notes, collections, and blob I/O for
    thumbnails / cached embeddings. Owning service: never calls search or graph.
servers:
  - url: http://deedlit-catalog:8000
paths:
...
  /images:
    post:
      summary: Create/upsert an image record (keyed by sha256)
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ImageUpsert" } } } }
      responses: { "200": { description: Upserted, content: { application/json: { schema: { $ref: "#/components/schemas/Image" } } } } }
    get:
      summary: List/search images
      parameters:
        - { name: tag, in: query, description: "Include tag (repeatable); image must carry ALL.", schema: { type: array, items: { type: string } } }
        - { name: exclude_tag, in: query, description: "Exclude tag (repeatable); image must carry NONE.", schema: { type: array, items: { type: string } } }
        - { name: favorite, in: query, schema: { type: boolean } }
        - { name: rating_gte, in: query, description: "Keep images rated at least this (0-5).", schema: { type: integer, minimum: 0, maximum: 5 } }
        - { name: safety, in: query, description: "Content-safety classes to include (repeatable); omit for all.", schema: { type: array, items: { type: string, enum: [sfw, nsfw, explicit] } } }
        - { name: path, in: query, description: "Keep images whose on-disk file path contains this fragment (separator-insensitive substring).", schema: { type: string } }
        - { name: sort, in: query, description: "Server-side ordering for the browse grid.", schema: { type: string, enum: [newest, oldest, rating_desc, rating_asc, name_asc, name_desc], default: newest } }
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses: { "200": { description: "OK — lightweight summaries (basic card fields only; no workflow/api-prompt graphs, params, negative, or description). Fetch the full record via GET /images/{sha256}.", content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/ImageSummary" } } } } } }
  /images/batch-delete:
    post:
      summary: Bulk hard-delete many image records (+ blobs); not the source files
      description: >
        The batch counterpart to DELETE /images/{sha256}: two set-based SQL deletes
        (asset refs, then the image rows — FK children cascade) plus per-blob
        object-store cleanup for exactly the records that existed. Returns the
        deleted + missing sha256s.
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { sha256s: { type: array, items: { $ref: "#/components/schemas/Sha256" }, minItems: 1, maxItems: 1000 } }, required: [sha256s] } } } }
      responses: { "200": { description: OK, content: { application/json: { schema: { type: object, properties: { status: { type: string }, deleted: { type: array, items: { $ref: "#/components/schemas/Sha256" } }, missing: { type: array, items: { $ref: "#/components/schemas/Sha256" } } } } } } } }
  /images/{sha256}:
    parameters: [ { name: sha256, in: path, required: true, schema: { $ref: "#/components/schemas/Sha256" } } ]
    get: { summary: Read one image, responses: { "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Image" } } } }, "404": { description: Not found } } }
    patch:
      summary: Update mutable fields (rating/favorite/tags)
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ImagePatch" } } } }
      responses: { "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Image" } } } } }
    delete:
      summary: Delete an image record (+ sha256-keyed blobs); not the source file
      description: >
        Hard-deletes the Postgres row (FK children cascade) plus the per-image
        asset references, then drops the sha256-keyed RustFS blobs (thumbnail +
        cached embedding). The original file on disk is never touched. Note /
        collection membership refs are left intact so a re-ingest re-attaches.
      responses:
        "200": { description: Deleted, content: { application/json: { schema: { type: object, properties: { status: { type: string } } } } } }
        "404": { description: Not found }
  /images/{sha256}/rating:
    put:
      parameters: [ { name: sha256, in: path, required: true, schema: { $ref: "#/components/schemas/Sha256" } } ]
      summary: Set rating
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { rating: { type: integer, minimum: 0, maximum: 5 } }, required: [rating] } } } }
      responses: { "200": { description: OK } }
  /images/{sha256}/favorite:
    put:
      parameters: [ { name: sha256, in: path, required: true, schema: { $ref: "#/components/schemas/Sha256" } } ]
      summary: Set favorite
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { favorite: { type: boolean } }, required: [favorite] } } } }
      responses: { "200": { description: OK } }
  /images/count:
    get:
      summary: Total images matching a filter set
      description: >
        The count of images matching the same filters as GET /images (no
        sort/paging). Lets a report/export tool size the work set before paging.
      parameters:
        - { name: tag, in: query, description: "Include tag (repeatable); image must carry ALL.", schema: { type: array, items: { type: string } } }
        - { name: exclude_tag, in: query, description: "Exclude tag (repeatable); image must carry NONE.", schema: { type: array, items: { type: string } } }
        - { name: favorite, in: query, schema: { type: boolean } }
        - { name: rating_gte, in: query, schema: { type: integer, minimum: 0, maximum: 5 } }
        - { name: safety, in: query, description: "Content-safety classes to include (repeatable); omit for all.", schema: { type: array, items: { type: string, enum: [sfw, nsfw, explicit] } } }
        - { name: path, in: query, description: "Keep images whose on-disk file path contains this fragment (separator-insensitive substring).", schema: { type: string } }
      responses: { "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/CountResult" } } } } }
  /images/unlabeled:
    get:
      summary: sha256 of images missing an AI label/description
      description: >
        The label-backfill work set: cataloged images with no
        deedlit.labelagent description. Returns { sha256: [...] }. Paged.
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 500 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses: { "200": { description: OK, content: { application/json: { schema: { type: object, properties: { sha256: { type: array, items: { $ref: "#/components/schemas/Sha256" } } }, required: [sha256] } } } } }
  /images/directories:
    get:
      summary: Distinct source directories with per-directory image counts
      description: >
        The distinct images.directory values with live-image totals, biggest
        folder first. Backs the library's split-by-source-directory section
        headers (true folder totals, not just the loaded page).
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 2000, minimum: 1, maximum: 10000 } }
      responses: { "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/DirectoryCount" } } } } } }
  /tags:
    get:
      summary: Tag-name autocomplete (names matching prefix, most-used first)
      parameters:
        - { name: prefix, in: query, schema: { type: string, default: "" } }
        - { name: limit, in: query, schema: { type: integer, default: 10, minimum: 1, maximum: 50 } }
      responses: { "200": { description: OK, content: { application/json: { schema: { type: array, items: { type: string } } } } } }
  /tags/report:
    get:
      summary: Full tag inventory with per-tag image counts, paged
      description: >
        Every tag plus how many distinct live images carry it (most-used first),
        with a `total` for paging the whole inventory. The report counterpart to
        GET /tags (which is a capped autocomplete of bare names).
      parameters:
        - { name: prefix, in: query, schema: { type: string, default: "" } }
        - { name: limit, in: query, schema: { type: integer, default: 200, minimum: 1, maximum: 1000 } }
        - { name: offset, in: query, schema: { type: integer, default: 0, minimum: 0 } }
      responses: { "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/TagReport" } } } } }
...
  /blobs/{sha256}/{kind}:
    parameters:
      - { name: sha256, in: path, required: true, schema: { $ref: "#/components/schemas/Sha256" } }
      - { name: kind, in: path, required: true, schema: { type: string, enum: [thumbnail, embedding, sparse] } }
    get: { summary: Fetch a RustFS blob, responses: { "200": { description: Blob, content: { application/octet-stream: { schema: { type: string, format: binary } } } }, "404": { description: Not found } } }
    put:
      summary: Store a RustFS blob (sha256-keyed)
      requestBody: { required: true, content: { application/octet-stream: { schema: { type: string, format: binary } } } }
      responses: { "200": { description: Stored } }
  /notes:
    post:
      summary: Create a note (Editor.js blocks + ordered image refs)
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/NoteUpsert" } } } }
      responses: { "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Note" } } } } }
  /notes/{id}:
    parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
    get: { summary: Read note, responses: { "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Note" } } } } } }
    put: { summary: Update note, requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/NoteUpsert" } } } }, responses: { "200": { description: OK } } }
  /notes/by-image/{sha256}:
    get:
      parameters: [ { name: sha256, in: path, required: true, schema: { $ref: "#/components/schemas/Sha256" } } ]
      summary: Notes referencing an image
      responses: { "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Note" } } } } } }
  /collections:
    post:
      summary: Create a collection
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CollectionUpsert" } } } }
      responses: { "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Collection" } } } } }
    get: { summary: List collections, responses: { "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Collection" } } } } } } }
  /collections/{id}/images:
    put:
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      summary: Set/reorder collection membership (ordered sha256 list)
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { images: { type: array, items: { $ref: "#/components/schemas/Sha256" } } }, required: [images] } } } }
      responses: { "200": { description: OK } }
  /folders:
    post:
      summary: Register a source folder (idempotent on path)
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/SourceFolderUpsert" } } } }
      responses: { "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/SourceFolder" } } } } }
    get: { summary: List configured folders (+ derived coverage), responses: { "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/SourceFolder" } } } } } } }
  /folders/{id}:
    parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
    get: { summary: Read one folder, responses: { "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/SourceFolder" } } } }, "404": { description: Not found } } }
    patch:
      summary: Update folder controls + scan-state (partial)
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/SourceFolderPatch" } } } }
      responses: { "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/SourceFolder" } } } }, "404": { description: Not found } }
    delete:
      summary: Remove a folder from the registry (keeps its cataloged images)
      responses: { "200": { description: Deleted }, "404": { description: Not found } }
...
components:
  schemas:
...
    Sha256: { type: string, pattern: "^[a-f0-9]{64}$" }
    CountResult:
      type: object
      properties: { count: { type: integer } }
      required: [count]
    DirectoryCount:
      type: object
      description: "One source directory + how many live images live directly under it (split-by-source-directory section header)."
      properties:
        directory: { type: string }
        image_count: { type: integer }
      required: [directory, image_count]
...
    TagReport:
      type: object
      properties:
        total: { type: integer, description: "Tags matching the prefix (for paging)" }
        items: { type: array, items: { $ref: "#/components/schemas/TagCount" } }
      required: [total, items]
...
    ImageUpsert:
      type: object
      properties:
        sha256: { $ref: "#/components/schemas/Sha256" }
        filepath:
          type: [string, "null"]
          description: >-
            Original on-disk path of the source file captured at ingest, so an
            image is identifiable when inspecting the catalog (the id is the
            opaque sha256). Stored INSERT-only — a later reindex (which has no
            path) never overwrites it. Null when unknown.
        phash: { type: [string, "null"] }
        width: { type: [integer, "null"] }
        height: { type: [integer, "null"] }
        sourceTool: { type: [string, "null"] }
        prompt: { type: [string, "null"] }
        negative: { type: [string, "null"] }
        tags: { type: array, items: { type: string } }
        params: { $ref: "#/components/schemas/Params" }
        references: { type: array, items: { $ref: "#/components/schemas/AssetRef" } }
        workflow_json: { type: [object, "null"] }
        api_prompt_json: { type: [object, "null"] }
        safety:
          type: [string, "null"]
          enum: [sfw, nsfw, explicit, null]
          description: AI content-safety class (deedlit.labelagent); null = unclassified.
        description:
          type: [string, "null"]
          description: >-
            AI-generated image description (deedlit.labelagent). An expensive
            vision-LLM result, persisted in image_descriptions so it is
            retrievable/viewable without re-running the model. Re-derivable: a
            reindex without one keeps the stored value. Null = none.
      required: [sha256]
    Image:
      allOf:
        - { $ref: "#/components/schemas/ImageUpsert" }
        - type: object
          properties:
            rating: { type: [integer, "null"] }
            favorite: { type: boolean }
            created_at: { type: string, format: date-time }
    ImageSummary:
      description: >-
        Lightweight browse row returned by GET /images (one per grid card). A
        deliberate SUBSET of Image: only the fields a card renders. The heavy
        workflow_json / api_prompt_json ComfyUI graphs (tens-to-hundreds of KB
        each) plus the detail-only negative / params / description are OMITTED so
        a browse page (and its auto-refresh poll) stays small. Fetch the full
        record via GET /images/{sha256}.
      type: object
      properties:
        sha256: { $ref: "#/components/schemas/Sha256" }
        filepath: { type: [string, "null"] }
        directory: { type: [string, "null"], description: "Parent directory of filepath (separators normalized to '/') — the split-by-source-directory grouping key." }
        phash: { type: [string, "null"] }
        width: { type: [integer, "null"] }
        height: { type: [integer, "null"] }
        sourceTool: { type: [string, "null"] }
        prompt: { type: [string, "null"] }
        tags: { type: array, items: { type: string } }
        references: { type: array, items: { $ref: "#/components/schemas/AssetRef" } }
        rating: { type: [integer, "null"] }
        favorite: { type: boolean }
        safety: { type: [string, "null"], enum: [sfw, nsfw, explicit, null] }
        created_at: { type: [string, "null"], format: date-time }
        imported_at: { type: [string, "null"], format: date-time }
      required: [sha256]
    ImagePatch:
      type: object
      properties:
        rating: { type: [integer, "null"] }
        favorite: { type: [boolean, "null"] }
        tags: { type: [array, "null"], items: { type: string } }
        safety: { type: [string, "null"], enum: [sfw, nsfw, explicit, null] }
    NoteUpsert:
      type: object
      properties:
        title: { type: [string, "null"] }
        positive: { type: [string, "null"] }
        negative: { type: [string, "null"] }
        blocks: { type: object, description: "Editor.js block document", additionalProperties: true }
        imageRefs: { type: array, items: { $ref: "#/components/schemas/Sha256" }, description: "ordered" }
      required: [blocks]
...